Create VaultManager.sol
The contract inherit from the Ownable contract from OpenZeppelin, this means that it can only be created by the owner of the contract and only the owner can create new vault.

The contract uses a mapping to keep track of all the vaults that have been created and the balances of each vault's owner. The createVault function is used to create a new vault and assigns the msg.sender as the owner. The deposit function transfers tokens from the msg.sender to the vault and increases the msg.sender's balance in the vault. The withdraw function transfers tokens from the vault back to the msg.sender and decreases the msg.sender's balance in the vault. The function also validate if the vault exists and if the sender is the owner of the vault, also validate if the amount of the withdrawal is less or equal than the balance of the vault.

Note that this is a simplified example and there are many possible ways to implement a decentralized vault manager.